Include helpers in bindgen-tests crate
Put test generation behind a feature so we don't have to compile all
that for the 3ds target, just the host as a build depdendency.
