+++ test-runner/Cargo.toml
@@ -12,3 +12,4 @@ socket = []
 ctru-rs = { git = "https://github.com/rust3ds/ctru-rs" }
 ctru-sys = { git = "https://github.com/rust3ds/ctru-rs" }
 libc = { workspace = true }
+shim-3ds = { git = "https://github.com/rust3ds/shim-3ds.git" }
